use tests/src/utils.rs instead of common.rs
