+++ tests/tests/pbs_post_validators.rs
@@ -1,5 +1,3 @@
-mod common;
-
 use std::{
     sync::Arc,
     time::Duration,
@@ -14,7 +12,7 @@ use cb_pbs::{DefaultBuilderApi, PbsService, PbsState};
 use cb_tests::{
     mock_relay::{start_mock_relay_service, MockRelayState},
     mock_validator::MockValidator,
-    utils::{generate_mock_relay, setup_test_env},
+    utils::{generate_mock_relay, setup_test_env, to_pbs_config, get_pbs_static_config},
 };
 use eyre::Result;
 use tracing::info;
@@ -32,7 +30,7 @@ async fn test_register_validators() -> Result<()> {
     let mock_state = Arc::new(MockRelayState::new(chain, signer));
     tokio::spawn(start_mock_relay_service(mock_state.clone(), port + 1));
 
-    let config = common::to_pbs_config(chain, common::get_pbs_static_config(port), relays);
+    let config = to_pbs_config(chain, get_pbs_static_config(port), relays);
     let state = PbsState::new(config);
     tokio::spawn(PbsService::run::<(), DefaultBuilderApi>(state));
